first pass at operator_tools refactor
