+++ forest/benchmarking/operator_tools/__init__.py
@@ -1,3 +1,7 @@
-__all__ = ["apply_superoperator", "channel_approximation", "project_superoperators",
-           "random_operators", "superoperator_transformations", "validate_operator",
-           "validate_superoperator" ]
+from .apply_superoperator import *
+from .channel_approximation import *
+from .project_superoperators import *
+from .superoperator_transformations import *
+from .random_operators import *
+from .validate_operator import *
+from .validate_superoperator import *
