+++ forest/benchmarking/operator_tools/validate_operator.py
@@ -1,40 +1,8 @@
-"""A module allowing one to check if superoperators or channels are physical.
-
-We have arbitrarily decided to use a column stacking convention.
-
-For more information about the conventions used, look at the file in
-/docs/Superoperator representations.md
-
-Further references include:
-
-[GRAPTN] Tensor networks and graphical calculus for open quantum systems
-         Wood et al.
-         Quant. Inf. Comp. 15, 0579-0811 (2015)
-         (no DOI)
-         https://arxiv.org/abs/1111.6950
-
-[MATQO] On the Matrix Representation of Quantum Operations
-        Nambu et al.,
-        arXiv: 0504091 (2005)
-        https://arxiv.org/abs/quant-ph/0504091
-
-[DUAL] On duality between quantum maps and quantum states
-       Zyczkowski et al.,
-       Open Syst. Inf. Dyn. 11, 3 (2004)
-       https://dx.doi.org/10.1023/B:OPSY.0000024753.05661.c2
-       https://arxiv.org/abs/quant-ph/0401119
-
+"""A module allowing one to check properties of operators or matrices.
 """
-from typing import Sequence
 import numpy as np
-from forest.benchmarking.utils import partial_trace
-from forest.benchmarking.operator_tools.superoperator_transformations import choi2kraus
-from forest.benchmarking.operator_tools.superoperator_tools import apply_choi_matrix_2_state
 
 
-# ==================================================================================================
-# Check properties of operators or matrices
-# ==================================================================================================
 def is_square_matrix(matrix: np.ndarray) -> bool:
     """
     Checks if a matrix is square.
